backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
alembic==1.13.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
google-generativeai==0.8.3
python-docx==1.1.2
python-pptx==0.6.23
pymysql==1.1.1

